Suppress -Werror=shorten-64-to-32
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@39933 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ addr2line.c
@@ -731,7 +731,7 @@ ksprintn(char *nbuf, uintmax_t num, int base, int *lenp, int upper)
 		*++p = upper ? toupper(c) : c;
 	} while (num /= base);
 	if (lenp)
-		*lenp = p - nbuf;
+		*lenp = (int)(p - nbuf);
 	return (p);
 }
 
@@ -946,7 +946,7 @@ reswitch:	switch (ch = (u_char)*fmt++) {
 			if (p == NULL)
 				p = "(null)";
 			if (!dot)
-				n = strlen (p);
+				n = (int)strlen (p);
 			else
 				for (n = 0; n < dwidth && p[n]; n++)
 					continue;
